+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add dev only installs in #32 by @Stefanuk12
 - Add `PESDE_HOME` variable (defaults to $HOME/.pesde) to override pesde's directory's location by @daimond113
 - Add `--target` argument to execute command by @daimond113
+- Add `engines` when initializing a project based on the scripts package's by @daimond113
 
 ### Fixed
 - Download engines in install step rather than lazily by @daimond113